Move analysis into analyzeme
