+++ analyzeme/Cargo.toml
@@ -10,6 +10,7 @@ decodeme = { path = "../decodeme" }
 memchr = "2"
 measureme = { path = "../measureme" }
 rustc-hash = "1.0.1"
+serde = { version = "1.0", features = ["derive"] }
 
 # Depending on older versions of this crate allows us to keep supporting older
 # file formats.
