+++ analyzeme/src/lib.rs
@@ -10,6 +10,7 @@
 //! To retrieve an `Iterator` of all of the events in the file,
 //! call the [`ProfilingData::iter()`] method.
 
+mod analysis;
 mod file_formats;
 mod profiling_data;
 mod stack_collapse;
@@ -17,6 +18,7 @@ pub mod testing_common;
 
 pub use crate::profiling_data::{ProfilingData, ProfilingDataBuilder};
 pub use crate::stack_collapse::collapse_stacks;
+pub use analysis::{AnalysisResults, ArtifactSize, QueryData};
 pub use decodeme::event::Event;
 pub use decodeme::event_payload::{EventPayload, Timestamp};
 pub use decodeme::lightweight_event::LightweightEvent;
